+++ infra/ecr.tf
@@ -45,6 +45,8 @@ data "aws_iam_policy_document" "aws_vpc_endpoint_ecr" {
 
     actions = [
       "ecr:DescribeImages",
+      "ecr:BatchGetImage",
+      "ecr:PutImage",
     ]
 
     resources = [