+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Updated text on "user unauthorised to use tools" page to remove the requirement for them to have SC.
 
+### Added
+
+- Initial version of button to release a visualisation to production
+
 ## 2020-04-06
 
 ### Added